WaveformsAtAPointFT: change constructor interface and windowing.
Now constructor takes WindowBeginTime and WindowEndTime instead
of WindowNCycles.
WindowNCycles was problematic because the amount of windowing
depended on where the first zero crossing is located.  So if
you changed theta,phi (or some other quantity) continuously so that
the first data point in the waveform passed through zero continuously,
the result of WaveformAtAPointFT changed *discontinuously* at
this zero-crossing.  Now a windowing range is passed in, and this
can be provided externally.

To prevent old code from using the old interface (and implicitly
casting the integer WindowNCycles into a double WindowBeginTime), the angles
Vartheta,Varphi are now specified as a components of a vector instead
of two separate parameters.